pyproject.toml
[project]
name = "batch-cli"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "mmysior", email = "mysior.marek@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.50.0",
    "click>=8.1.8",
    "openai>=1.76.0",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]

[project.scripts]
batch-cli = "batch_cli.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
